requirements.txt
beautifulsoup4==4.14.3
ffmpeg-python==0.2.0
plyer==2.1.0
pywin32==311
requests==2.32.5
winshell==0.6
